Fix content type for store_url call.
For some reason (I have an email in to Filepicker to find out why),
the Filepicker API does not work with a content-type of
application/x-www-url-encoded-form. RestClient forces this
content-type when a hash is provided as a payload. This commit changes
the payload and specifies a content type of text/plain.